Create To do list website
To-do lists are a list of tasks that an individual needs to complete or accomplish.
To-Do List project is an application specially built to keep track of errands or tasks that need to be done. This application will be like a task keeper where the user would be able to enter the tasks that they need to do. Once they are done with their tasks they can also remove them from the list.